+++ lib/composables/use-months-list.ts
@@ -8,7 +8,8 @@ interface MonthsListOptions {
   format?: MonthInputFormat;
 }
 
-export function useMonthsList ({ locale }: NormalizedCalendarOptions): (options?: MonthsListOptions) => string[] {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function useMonthsList ({ locale }: NormalizedCalendarOptions<any, any>): (options?: MonthsListOptions) => string[] {
   return ({ format }: MonthsListOptions = {}): string[] => {
     const anyJanuary = new Date(2000, 0, 1);
     const months = Array.from(Array(NB_OF_MONTHS).keys()).map(i => addMonths(anyJanuary, i));